0.0.5 | added remaining of the entities, service and converters

+++ src/main/java/model/dao/UserDAO.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 import model.entity.User;
 
 public interface UserDAO {
-	Optional<User> getUser(int id);
+	Optional<User> getUser(Long id);
 	Optional<User> getUser(String username);
-    int addUser(User user);
+	Long addUser(User user);
     void deleteUser(String id);
     void updateUser(User user);
     Optional<User> getUser(String login, String pass, boolean isAdmin);
